Markdown tests to use their own project too, with sample markdown.
