--- src/Spectre.Console.Tests/Expectations/MarkdownTests.SimpleHeaderAndTextDoc.verified.txt
@@ -1,11 +1,0 @@
-─────────────────────────────── Heading level 1 ────────────────────────────────
-Some text
-─────────────────────────────── Heading level 2 ────────────────────────────────
- - a list
- - of things
-                                Heading level 3                                 
-                                Heading level 4                                 
-                                Heading level 5                                 
-Some text
-                                Heading level 6                                 
-
